move one folder down
