+++ src/json-reporter.js
